Fix note event signaling
Consolidate the concept of note/gate so that we can properly associate
noteoff events with the note that is turning off.

Approach is to have a "note event" integer signal, where >=0 indicates a
noteon event with that MIDI note value, and <0 indicates a noteoff event
with the negated value of the signal as the MIDI note value.