(poly) "fix" 6-note poly bug
As it turns out, typical USB keyboards only support holding down 6 keys
at a time. This was leading to weird issues in poly mode when allocating
voces > 6, or subtle oddities with voice stealing when voices < 6.

The workaround is to limit polyphony to 6 voices, which will also have
the side effect of making voice stealing not a thing in poly mode when
performing on a computer keyboard. It also works nicely for
pseudo-voices in mono mode, since you won't be able to attempt to add
another pseudo voice past the point that the keyboard stops supporting
it.